Do not run hooks if plugin not included in Gemfile
Including the `plugin` directive in a Gemfile makes Bundler install that
plugin but, unlike the `gem` directive, removing it does not
subsequently remove the plugin or its hooks so the hooks will continue
to run. I think this is not very friendly behaviour so now we parse the
Gemfile (using the same strategy as Bundler itself uses) and verify that
the plugin is explicitly included. If not, we don't run our hook and we
act as though we don't exist.