No longer officially a Bundler plugin
Unfortunately the Bundler plugin system is not yet ready for use and has
too many problems so we are resorting to providing an executable which
can be run manually. Although this means we can no longer enforce a
policy by hooking into `bundle install` we can still leave it to the
user's discretion to include in any CI build scripts.

bundle exec strong_versions